solar-ml/requirements.txt
streamlit
influxdb-client
scikit-learn
joblib
python-dotenv
langchain
langchain-openai
openai
